requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pillow==10.1.0
pandas==2.1.3
tkinterdnd2==0.3.0
matplotlib==3.8.2
